# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventCreate": {
        "additionalProperties": false,
        "description": "Manual event entry. Stored with ``source='manual'`` and ``verified=false`` until a\nreviewer verifies it; without evidence the UI shows it as uncited.",
        "properties": {
          "cause_text": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause Text"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "event_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "event_type": {
            "enum": [
              "LOSS",
              "KICK",
              "STUCK",
              "TIGHT",
              "TORQUE",
              "INSTAB",
              "BALLING",
              "OVERP",
              "GAS",
              "CEMENT",
              "CASING",
              "FISH",
              "EQUIP",
              "WAIT",
              "OTHER_NPT"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRefIn"
            },
            "maxItems": 50,
            "title": "Evidence",
            "type": "array"
          },
          "formation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Id"
          },
          "hole_size_in": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 40.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hole Size In"
          },
          "md_m": {
            "anyOf": [
              {
                "maximum": 15000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md M"
          },
          "mitigations": {
            "items": {
              "$ref": "#/components/schemas/MitigationCreate"
            },
            "maxItems": 20,
            "title": "Mitigations",
            "type": "array"
          },
          "mw_sg": {
            "anyOf": [
              {
                "maximum": 2.6,
                "minimum": 0.8,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mw Sg"
          },
          "npt_hours": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npt Hours"
          },
          "params": {
            "$ref": "#/components/schemas/EventParamsIn"
          },
          "resolved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved"
          },
          "severity": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "subtype": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "t_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T End"
          },
          "t_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T Start"
          },
          "well_id": {
            "title": "Well Id",
            "type": "integer"
          },
          "wellbore_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wellbore Id"
          }
        },
        "required": [
          "well_id",
          "event_type"
        ],
        "title": "EventCreate",
        "type": "object"
      },
      "EventDetail": {
        "properties": {
          "cause_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cause Text"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "event_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "event_type": {
            "enum": [
              "LOSS",
              "KICK",
              "STUCK",
              "TIGHT",
              "TORQUE",
              "INSTAB",
              "BALLING",
              "OVERP",
              "GAS",
              "CEMENT",
              "CASING",
              "FISH",
              "EQUIP",
              "WAIT",
              "OTHER_NPT"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence",
            "type": "array"
          },
          "formation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Formation name at the event depth",
            "title": "Formation"
          },
          "formation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formation Id"
          },
          "hole_size_in": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hole Size In"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "lesson_card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LessonCardContent"
              },
              {
                "type": "null"
              }
            ]
          },
          "md_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md M"
          },
          "mitigations": {
            "items": {
              "$ref": "#/components/schemas/MitigationOut"
            },
            "title": "Mitigations",
            "type": "array"
          },
          "mw_sg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mw Sg"
          },
          "npt_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Npt Hours"
          },
          "params": {
            "$ref": "#/components/schemas/EventParams"
          },
          "resolved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved"
          },
          "severity": {
            "anyOf": [
              {
                "enum": [
                  "low",
                  "medium",
                  "high"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "source": {
            "enum": [
              "rules",
              "llm",
              "manual",
              "import"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "subtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtype"
          },
          "synthetic": {
            "description": "True when the well belongs to a SYNTHETIC field",
            "title": "Synthetic",
            "type": "boolean"
          },
          "t_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T End"
          },
          "t_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T Start"
          },
          "tvd_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvd M"
          },
          "tvdss_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tvdss M"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          },
          "well_id": {
            "title": "Well Id",
            "type": "integer"
          },
          "well_name": {
            "title": "Well Name",
            "type": "string"
          },
          "wellbore_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wellbore Id"
          }
        },
        "required": [
          "confidence",
          "verified",
          "evidence",
          "id",
          "well_id",
          "well_name",
          "wellbore_id",
          "synthetic",
          "event_type",
          "subtype",
          "severity",
          "event_date",
          "t_start",
          "t_end",
          "md_m",
          "tvd_m",
          "tvdss_m",
          "formation_id",
          "formation",
          "hole_size_in",
          "mw_sg",
          "npt_hours",
          "resolved",
          "source",
          "status",
          "params",
          "cause_text",
          "description",
          "lesson_card",
          "mitigations",
          "verified_by",
          "verified_at",
          "created_at",
          "updated_at"
        ],
        "title": "EventDetail",
        "type": "object"
      },
...
      "EventPage": {
        "description": "Cursor page: pass ``next_cursor`` back as ``cursor``; null means the last page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "EventVerify": {
        "additionalProperties": false,
        "description": "Verify (or un-verify) an event, or reject it as not a real event.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "rejected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set 'rejected' to hide a false extraction; null = unchanged",
            "title": "Status"
          },
          "verified": {
            "default": true,
            "title": "Verified",
            "type": "boolean"
          }
        },
        "title": "EventVerify",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "eRTMAC-NWIS offset-well knowledge and decision support (SIH PS 121)",
    "title": "SMRITI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "description": "Active events matching every given filter, ordered by (well name, md_m, id).\n\n``formation`` matches a formation name or synonym (case-insensitive). With ``radius_km``\nthe events of ``well_id`` *and* of every well whose surface location lies within the\nradius are returned; without it, only ``well_id``'s. Rejected events are never listed.",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "description": "Repeat to match any of several types",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "LOSS",
                      "KICK",
                      "STUCK",
                      "TIGHT",
                      "TORQUE",
                      "INSTAB",
                      "BALLING",
                      "OVERP",
                      "GAS",
                      "CEMENT",
                      "CASING",
                      "FISH",
                      "EQUIP",
                      "WAIT",
                      "OTHER_NPT"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat to match any of several types",
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "formation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Formation"
            }
          },
          {
            "in": "query",
            "name": "well_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Well Id"
            }
          },
          {
            "description": "Radius around well_id's surface location (km)",
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 100,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Radius around well_id's surface location (km)",
              "title": "Radius Km"
            }
          },
          {
            "in": "query",
            "name": "tvdss_from_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tvdss From M"
            }
          },
          {
            "in": "query",
            "name": "tvdss_to_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tvdss To M"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "description": "Only verified / unverified",
            "in": "query",
            "name": "verified",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only verified / unverified",
              "title": "Verified"
            }
          },
          {
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from the previous page's next_cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from the previous page's next_cursor",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search drilling events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Stores the event with ``source='manual'``, ``verified=false``; depth references\n(``tvd_m``/``tvdss_m``) and the formation are derived from ``md_m`` when not given.\n404 if the well does not exist; 422 if evidence spans are not on the cited page.",
        "operationId": "create_event_api_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record an event manually",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "description": "One event (active or rejected) with its mitigations in the order tried.",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event detail with mitigations, evidence and lesson card",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}/verify": {
      "patch": {
        "description": "Sets ``verified``/``verified_by``/``verified_at`` (and ``status`` when given) and\ncloses any pending review items for the event.",
        "operationId": "verify_event_api_v1_events__event_id__verify_patch",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventVerify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify, un-verify or reject an event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/wells/{well_id}/events/timeline": {
      "get": {
        "description": "Active events of the well plus its DDR NPT lines, for the Well 360 Events tab.",
        "operationId": "get_event_timeline_api_v1_wells__well_id__events_timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "well_id",
            "required": true,
            "schema": {
              "title": "Well Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventTimeline"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chronological event list for one well",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}